create default base plugin functionality // create stub plugin for users // add docs
